343 : Add step 3 component
