+++ src/components/simulation/ExplanationSelectionObstacles.vue
@@ -2,10 +2,7 @@
 import step2 from '@/assets/illustrations/Steps.jpg'
 import validation from '@/assets/icons/interface-validation.svg'
 import deleted from '@/assets/icons/interface-delete.svg'
-import economies from '@/assets/icons/economies.svg'
-import { useSimulationStore } from '@/stores/simulations'
-
-const simulationStore = useSimulationStore()
+import BoxStep3 from '@/components/simulation/BoxStep3.vue'
 
 const props = defineProps({
   step: {
@@ -17,11 +14,6 @@ const props = defineProps({
     required: true,
   },
 })
-
-function goToThirdStep() {
-  simulationStore.setCurrentSubStep(1)
-  simulationStore.goToNextStep()
-}
 </script>
 
 <template>
@@ -52,28 +44,6 @@ function goToThirdStep() {
   </div>
   <div v-if="props.step === 2 && props.subStep === 2">
     <div></div>
-    <div
-      class="flex flex-col justify-center text-center py-6 px-6 gap-4 box-border border-[1px] border-slate-600 rounded-lg"
-    >
-      <div class="gap-px">
-        <p class="font-normal text-xs font-dm-sans text-neutral-500">Étape 3</p>
-        <img
-          class="w-[88px] h-[88px] max-w-none mx-auto"
-          :src="economies"
-          alt=""
-        />
-        <h4 class="font-dm-sans text-[18px] leading-6 font-bold">
-          Quelles économies d'énergie <br />grâce à cette installation ?
-        </h4>
-      </div>
-      <button
-        class="bg-black shadow-sm rounded-lg gap-3 px-4 py-3 items-center flex flex-row justify-center text-white"
-        @click="goToThirdStep()"
-      >
-        <p class="font-semibold text-base font-dm-sans">
-          Calculer mes économies d'énergie
-        </p>
-      </button>
-    </div>
+    <BoxStep3></BoxStep3>
   </div>
 </template>
