+++ src/views/StepSunshineView.vue
@@ -7,7 +7,7 @@ import { useSimulationStore } from '@/stores/simulations'
 import SimulationSteps from '@/components/simulation/SimulationSteps.vue'
 import StepTitle from '@/components/simulation/StepTitle.vue'
 import StepDescription from '@/components/simulation/StepDescription.vue'
-import CustomContent from '@/components/simulation/CustomContent.vue'
+import ExplanationSelectionObstacles from '@/components/simulation/ExplanationSelectionObstacles.vue'
 import FooterButtons from '@/components/simulation/FooterButtons.vue'
 import RoofAccordionOptions from '@/components/simulation/RoofAccordionOptions.vue'
 
@@ -31,10 +31,10 @@ onBeforeMount(() => {
     :step="simulationStore.currentStep"
     :substep="simulationStore.currentSubStep"
   ></StepDescription>
-  <CustomContent
+  <ExplanationSelectionObstacles
     :step="simulationStore.currentStep"
     :subStep="simulationStore.currentSubStep"
-  ></CustomContent>
+  ></ExplanationSelectionObstacles>
   <RoofAccordionOptions
     v-if="simulationStore.currentStep === 1"
   ></RoofAccordionOptions>
